+++ lib/src/producer/platform.dart
@@ -110,7 +110,7 @@ extension on String {
 }
 
 extension on File {
-  /// Write the content of the the settings.gradle.kts of a Klutter plugin.
+  /// Write the content of the settings.gradle.kts of a Klutter plugin.
   void writeSettingsGradleContent(String pluginName) {
     writeAsStringSync('''
             // Copyright (c) 2021 - 2022 Buijs Software
@@ -138,7 +138,7 @@ extension on File {
         .format);
   }
 
-  /// Write the content of the the settings.gradle.kts of a Klutter plugin.
+  /// Write the content of the build.gradle.kts of a Klutter plugin.
   void writeRootBuildGradleContent(String pluginName) {
     writeAsStringSync('''
           buildscript {
@@ -151,7 +151,7 @@ extension on File {
           |    dependencies {
           |        classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.10")
           |        classpath("com.android.tools.build:gradle:7.0.4")
-          |        classpath("dev.buijs.klutter:core:$klutterGradleVersion")
+          |        classpath("dev.buijs.klutter:kore:$klutterGradleVersion")
           |        classpath("dev.buijs.klutter.gradle:dev.buijs.klutter.gradle.gradle.plugin:$klutterGradleVersion")
           |    }
           |}
@@ -174,25 +174,8 @@ extension on File {
           |    }
           |
           |}
-          |
-          |tasks.register("klutterInstallPlatform", Exec::class) {
-          |    commandLine("bash", "./gradlew", "clean", "build", "-p", "platform")
-          |    finalizedBy("klutterCopyAarFile", "klutterCopyFramework")
-          |}
-          |
-          |tasks.register("klutterCopyAarFile", Copy::class) {
-          |    from("platform/build/outputs/aar/$pluginName-release.aar")
-          |    into("android/klutter")
-          |    rename { fileName ->
-          |        fileName.replace("$pluginName-release", "platform")
-          |    }
-          |}
-          |
-          |tasks.register("klutterCopyFramework", Copy::class) {
-          |    from("platform/build/fat-framework/release")
-          |    into("ios/Klutter")
-          |}'''
-        .format);
+          |'''.format
+    );
   }
 
   /// Write the content of gradle.properties.
@@ -301,12 +284,18 @@ class PlatformModule {
       |    plugin { 
       |       name = "$pluginName"
       |    }
+      |
+      |    dependencies {
+      |       implementation("annotations")
+      |    } 
+      |
       |}
       |
       |kotlin {
       |    android()
       |    iosX64()
       |    iosArm64()
+      |    iosSimulatorArm64()
       |
       |    cocoapods {
       |        summary = "Some description for the Shared Module"
@@ -321,8 +310,7 @@ class PlatformModule {
       |
       |        val commonMain by getting {
       |            dependencies {
-      |                api("org.jetbrains.kotlinx:kotlinx-serialization-json:1.3.2")
-      |                api("dev.buijs.klutter:annotations-kmp:$klutterGradleVersion")
+      |                implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.3.2")
       |            }
       |        }
       |
@@ -351,19 +339,23 @@ class PlatformModule {
       |
       |        val iosX64Main by getting
       |        val iosArm64Main by getting
+      |        val iosSimulatorArm64Main by getting
       |        val iosMain by creating {
       |            dependsOn(commonMain)
       |            iosX64Main.dependsOn(this)
       |            iosArm64Main.dependsOn(this)
+      |            iosSimulatorArm64Main.dependsOn(this)
       |            dependencies {}
       |        }
       |
       |        val iosX64Test by getting
       |        val iosArm64Test by getting
+      |        val iosSimulatorArm64Test by getting
       |        val iosTest by creating {
       |            dependsOn(commonTest)
       |            iosX64Test.dependsOn(this)
       |            iosArm64Test.dependsOn(this)
+      |            iosSimulatorArm64Test.dependsOn(this)
       |        }
       |    }
       |}
@@ -420,7 +412,7 @@ class PlatformModule {
       ..writeAsStringSync("""
       package $packageName.platform
       |
-      |import dev.buijs.klutter.annotations.kmp.KlutterAdaptee
+      |import dev.buijs.klutter.annotations.KlutterAdaptee
       |
       |class Greeting {
       |