+++ test/src/producer/project_test.dart
@@ -28,16 +28,6 @@ void main() {
 
   const pluginName = "impressive_dependency";
 
-  test("Verify exception is thrown if example folder does not exist", () {
-    expect(() => writeExampleMainDartFile(
-      pathToExample: "fake",
-      pluginName: pluginName,
-    ), throwsA(predicate((e) =>
-        e is KlutterException &&
-        e.cause.startsWith("Path does not exist:") &&
-        e.cause.endsWith("/fake"))));
-  });
-
   test("Verify example/lib/main.dart is created if it does not exist", () {
 
     final example = Directory("${Directory.systemTemp.path}/wem1".normalize)
@@ -53,29 +43,29 @@ void main() {
     expect(mainDart.readAsStringSync().replaceAll(" ", ""), r"""
             import 'package:flutter/material.dart';
             import 'dart:async';
-            
+
             import 'package:impressive_dependency/impressive_dependency.dart';
-            
+
             void main() {
               runApp(const MyApp());
             }
-            
+
             class MyApp extends StatefulWidget {
               const MyApp({Key? key}) : super(key: key);
-            
+
               @override
               State<MyApp> createState() => _MyAppState();
             }
-            
+
             class _MyAppState extends State<MyApp> {
               String _platformVersion = 'Unknown';
-            
+
               @override
               void initState() {
                 super.initState();
                 initPlatformState();
               }
-            
+
               // Platform messages are asynchronous, so we initialize in an async method.
               Future<void> initPlatformState() async {
                 // Klutter generated Adapters don't throw exceptions but always return a
@@ -84,18 +74,18 @@ void main() {
                   String platformVersion = response.isSuccess()
                       ? response.object
                       : response.exception.toString();
-            
+
                   // If the widget was removed from the tree while the asynchronous platform
                   // message was in flight, we want to discard the reply rather than calling
                   // setState to update our non-existent appearance.
                   if (!mounted) return;
-            
+
                   setState(() {
                     _platformVersion = platformVersion;
                   });
                 });
               }
-            
+
               @override
               Widget build(BuildContext context) {
                 return MaterialApp(